+++ examples/data-sources/nxos_isis_address_family/data-source.tf
@@ -1,0 +1,5 @@
+data "nxos_isis_address_family" "example" {
+  instance_name  = "ISIS1"
+  vrf            = "default"
+  address_family = "v4"
+}