+++ examples/data-sources/nxos_isis_overload/data-source.tf
@@ -1,0 +1,4 @@
+data "nxos_isis_overload" "example" {
+  instance_name = "ISIS1"
+  vrf           = "default"
+}