+++ examples/resources/nxos_isis_overload/resource.tf
@@ -1,0 +1,5 @@
+resource "nxos_isis_overload" "example" {
+  instance_name = "ISIS1"
+  vrf           = "default"
+  startup_time  = 60
+}