+++ internal/provider/resource_nxos_isis_address_family.go
@@ -24,15 +24,12 @@ import (
 	"fmt"
 
 	"github.com/CiscoDevNet/terraform-provider-nxos/internal/provider/helpers"
-	"github.com/hashicorp/terraform-plugin-framework-validators/int64validator"
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/booldefault"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/int64default"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringdefault"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
@@ -41,25 +38,25 @@ import (
 )
 
 // Ensure provider defined types fully satisfy framework interfaces
-var _ resource.Resource = &ISISVRFResource{}
-var _ resource.ResourceWithImportState = &ISISVRFResource{}
+var _ resource.Resource = &ISISAddressFamilyResource{}
+var _ resource.ResourceWithImportState = &ISISAddressFamilyResource{}
 
-func NewISISVRFResource() resource.Resource {
-	return &ISISVRFResource{}
+func NewISISAddressFamilyResource() resource.Resource {
+	return &ISISAddressFamilyResource{}
 }
 
-type ISISVRFResource struct {
+type ISISAddressFamilyResource struct {
 	clients map[string]*nxos.Client
 }
 
-func (r *ISISVRFResource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
-	resp.TypeName = req.ProviderTypeName + "_isis_vrf"
+func (r *ISISAddressFamilyResource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_isis_address_family"
 }
 
-func (r *ISISVRFResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
+func (r *ISISAddressFamilyResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		// This description is used by the documentation generator and the language server.
-		MarkdownDescription: helpers.NewResourceDescription("This resource can manage the IS-IS VRF configuration.", "isisDom", "Routing%20and%20Forwarding/isis:Dom/").AddParents("isis_instance").AddChildren("isis_address_family", "isis_overload").AddReferences("vrf").String,
+		MarkdownDescription: helpers.NewResourceDescription("This resource can manage the IS-IS Address Family configuration.", "isisDomAf", "Routing%20and%20Forwarding/isis:DomAf/").AddParents("isis_vrf").AddReferences("vrf").String,
 
 		Attributes: map[string]schema.Attribute{
 			"device": schema.StringAttribute{
@@ -80,123 +77,52 @@ func (r *ISISVRFResource) Schema(ctx context.Context, req resource.SchemaRequest
 					stringplanmodifier.RequiresReplace(),
 				},
 			},
-			"name": schema.StringAttribute{
+			"vrf": schema.StringAttribute{
 				MarkdownDescription: helpers.NewAttributeDescription("VRF name.").String,
 				Required:            true,
 				PlanModifiers: []planmodifier.String{
 					stringplanmodifier.RequiresReplace(),
 				},
 			},
-			"admin_state": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Administrative state.").AddStringEnumDescription("enabled", "disabled").AddDefaultValueDescription("enabled").String,
-				Optional:            true,
-				Computed:            true,
-				Default:             stringdefault.StaticString("enabled"),
+			"address_family": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Address family type.").AddStringEnumDescription("v4", "v6").AddDefaultValueDescription("v4").String,
+				Required:            true,
 				Validators: []validator.String{
-					stringvalidator.OneOf("enabled", "disabled"),
+					stringvalidator.OneOf("v4", "v6"),
+				},
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
 				},
 			},
-			"authentication_check_l1": schema.BoolAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Authentication Check for ISIS on Level-1.").AddDefaultValueDescription("true").String,
+			"segment_routing_mpls": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Segment routing for MPLS	").AddDefaultValueDescription("false").String,
 				Optional:            true,
 				Computed:            true,
-				Default:             booldefault.StaticBool(true),
+				Default:             booldefault.StaticBool(false),
 			},
-			"authentication_check_l2": schema.BoolAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Authentication Check for ISIS on Level-2.").AddDefaultValueDescription("true").String,
+			"enable_bfd": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Enabling BFD on all ISIS domain interfaces.").AddDefaultValueDescription("false").String,
 				Optional:            true,
 				Computed:            true,
-				Default:             booldefault.StaticBool(true),
+				Default:             booldefault.StaticBool(false),
 			},
-			"authentication_key_l1": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Authentication Key for IS-IS on Level-1.").String,
-				Optional:            true,
-			},
-			"authentication_key_l2": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Authentication Key for IS-IS on Level-2.").String,
-				Optional:            true,
-			},
-			"authentication_type_l1": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("IS-IS Authentication-Type for Level-1.").AddStringEnumDescription("clear", "md5", "unknown").AddDefaultValueDescription("unknown").String,
+			"prefix_advertise_passive_l1": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Prefix advertise passive only for level-1").AddDefaultValueDescription("false").String,
 				Optional:            true,
 				Computed:            true,
-				Default:             stringdefault.StaticString("unknown"),
-				Validators: []validator.String{
-					stringvalidator.OneOf("clear", "md5", "unknown"),
-				},
+				Default:             booldefault.StaticBool(false),
 			},
-			"authentication_type_l2": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("IS-IS Authentication-Type for Level-2.").AddStringEnumDescription("clear", "md5", "unknown").AddDefaultValueDescription("unknown").String,
+			"prefix_advertise_passive_l2": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Prefix advertise passive only level-2").AddDefaultValueDescription("false").String,
 				Optional:            true,
 				Computed:            true,
-				Default:             stringdefault.StaticString("unknown"),
-				Validators: []validator.String{
-					stringvalidator.OneOf("clear", "md5", "unknown"),
-				},
-			},
-			"bandwidth_reference": schema.Int64Attribute{
-				MarkdownDescription: helpers.NewAttributeDescription("The IS-IS domain bandwidth reference. This sets the default reference bandwidth used for calculating the IS-IS cost metric.").AddIntegerRangeDescription(0, 4294967295).AddDefaultValueDescription("40000").String,
-				Optional:            true,
-				Computed:            true,
-				Default:             int64default.StaticInt64(40000),
-				Validators: []validator.Int64{
-					int64validator.Between(0, 4294967295),
-				},
-			},
-			"banwidth_reference_unit": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Bandwidth reference unit.").AddStringEnumDescription("mbps", "gbps").AddDefaultValueDescription("mbps").String,
-				Optional:            true,
-				Computed:            true,
-				Default:             stringdefault.StaticString("mbps"),
-				Validators: []validator.String{
-					stringvalidator.OneOf("mbps", "gbps"),
-				},
-			},
-			"is_type": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("IS-IS domain type.").AddStringEnumDescription("l1", "l2", "l12").AddDefaultValueDescription("l12").String,
-				Optional:            true,
-				Computed:            true,
-				Default:             stringdefault.StaticString("l12"),
-				Validators: []validator.String{
-					stringvalidator.OneOf("l1", "l2", "l12"),
-				},
-			},
-			"metric_type": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("IS-IS metric type.").AddStringEnumDescription("narrow", "wide", "transition").AddDefaultValueDescription("wide").String,
-				Optional:            true,
-				Computed:            true,
-				Default:             stringdefault.StaticString("wide"),
-				Validators: []validator.String{
-					stringvalidator.OneOf("narrow", "wide", "transition"),
-				},
-			},
-			"mtu": schema.Int64Attribute{
-				MarkdownDescription: helpers.NewAttributeDescription("The configuration of link-state packet (LSP) maximum transmission units (MTU) is supported. You can enable up to 4352 bytes.").AddIntegerRangeDescription(256, 4352).AddDefaultValueDescription("1492").String,
-				Optional:            true,
-				Computed:            true,
-				Default:             int64default.StaticInt64(1492),
-				Validators: []validator.Int64{
-					int64validator.Between(256, 4352),
-				},
-			},
-			"net": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Holds IS-IS domain NET (address) value.").String,
-				Optional:            true,
-			},
-			"passive_default": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("IS-IS Domain passive-interface default level.").AddStringEnumDescription("l1", "l2", "l12", "unknown").AddDefaultValueDescription("unknown").String,
-				Optional:            true,
-				Computed:            true,
-				Default:             stringdefault.StaticString("unknown"),
-				Validators: []validator.String{
-					stringvalidator.OneOf("l1", "l2", "l12", "unknown"),
-				},
+				Default:             booldefault.StaticBool(false),
 			},
 		},
 	}
 }
 
-func (r *ISISVRFResource) Configure(ctx context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
+func (r *ISISAddressFamilyResource) Configure(ctx context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
 	// Prevent panic if the provider has not been configured.
 	if req.ProviderData == nil {
 		return
@@ -205,8 +131,8 @@ func (r *ISISVRFResource) Configure(ctx context.Context, req resource.ConfigureR
 	r.clients = req.ProviderData.(map[string]*nxos.Client)
 }
 
-func (r *ISISVRFResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
-	var plan ISISVRF
+func (r *ISISAddressFamilyResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
+	var plan ISISAddressFamily
 
 	// Read plan
 	diags := req.Plan.Get(ctx, &plan)
@@ -241,8 +167,8 @@ func (r *ISISVRFResource) Create(ctx context.Context, req resource.CreateRequest
 	helpers.SetFlagImporting(ctx, false, resp.Private, &resp.Diagnostics)
 }
 
-func (r *ISISVRFResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
-	var state ISISVRF
+func (r *ISISAddressFamilyResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
+	var state ISISAddressFamily
 
 	// Read state
 	diags := req.State.Get(ctx, &state)
@@ -283,8 +209,8 @@ func (r *ISISVRFResource) Read(ctx context.Context, req resource.ReadRequest, re
 	helpers.SetFlagImporting(ctx, false, resp.Private, &resp.Diagnostics)
 }
 
-func (r *ISISVRFResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
-	var plan ISISVRF
+func (r *ISISAddressFamilyResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
+	var plan ISISAddressFamily
 
 	// Read plan
 	diags := req.Plan.Get(ctx, &plan)
@@ -315,8 +241,8 @@ func (r *ISISVRFResource) Update(ctx context.Context, req resource.UpdateRequest
 	resp.Diagnostics.Append(diags...)
 }
 
-func (r *ISISVRFResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
-	var state ISISVRF
+func (r *ISISAddressFamilyResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+	var state ISISAddressFamily
 
 	// Read state
 	diags := req.State.Get(ctx, &state)
@@ -358,7 +284,7 @@ func (r *ISISVRFResource) Delete(ctx context.Context, req resource.DeleteRequest
 	resp.State.RemoveResource(ctx)
 }
 
-func (r *ISISVRFResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
+func (r *ISISAddressFamilyResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
 	resource.ImportStatePassthroughID(ctx, path.Root("id"), req, resp)
 
 	helpers.SetFlagImporting(ctx, true, resp.Private, &resp.Diagnostics)